chore(clients/go): moved commands into pkg
