+++ clients/go/Makefile
@@ -5,7 +5,7 @@ install-deps:
 
 .PHONY: test
 test:
-	go test -v ./commands/
+	go test -v ./pkg/commands/
 	go test -v ./pkg/entities/
 	go test -v ./worker/
 	go test -v ./tests/
