+++ clients/go/cmd/zbctl/internal/commands/createInstance.go
@@ -14,7 +14,7 @@
 package commands
 
 import (
-	"github.com/zeebe-io/zeebe/clients/go/commands"
+	"github.com/zeebe-io/zeebe/clients/go/pkg/commands"
 	"strings"
 
 	"github.com/spf13/cobra"
