+++ clients/go/cmd/zbctl/internal/commands/createWorker.go
@@ -18,7 +18,7 @@ import (
 	"bytes"
 	"fmt"
 	"github.com/spf13/cobra"
-	"github.com/zeebe-io/zeebe/clients/go/commands"
+	"github.com/zeebe-io/zeebe/clients/go/pkg/commands"
 	"github.com/zeebe-io/zeebe/clients/go/pkg/entities"
 	"github.com/zeebe-io/zeebe/clients/go/worker"
 	"github.com/zeebe-io/zeebe/clients/go/zbc"
