+++ clients/go/pkg/commands/activateJobs_command_test.go
@@ -15,15 +15,15 @@
 package commands
 
 import (
-    "github.com/golang/mock/gomock"
-    "github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
-    "github.com/zeebe-io/zeebe/clients/go/internal/pb"
-    "github.com/zeebe-io/zeebe/clients/go/pkg/entities"
-    "github.com/zeebe-io/zeebe/clients/go/utils"
-    "io"
-    "reflect"
-    "testing"
-    "time"
+	"github.com/golang/mock/gomock"
+	"github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
+	"github.com/zeebe-io/zeebe/clients/go/internal/pb"
+	"github.com/zeebe-io/zeebe/clients/go/pkg/entities"
+	"github.com/zeebe-io/zeebe/clients/go/utils"
+	"io"
+	"reflect"
+	"testing"
+	"time"
 )
 
 func TestActivateJobsCommand(t *testing.T) {
