+++ clients/go/pkg/commands/cancelInstance_command.go
