+++ clients/go/pkg/commands/cancelInstance_command_test.go
