+++ clients/go/pkg/commands/completeJob_command.go
