+++ clients/go/pkg/commands/completeJob_command_test.go
