+++ clients/go/pkg/commands/deploy_command.go
