+++ clients/go/pkg/commands/failJob_command_test.go
