+++ clients/go/pkg/commands/publishMessage_command.go
