+++ clients/go/pkg/commands/publishMessage_command_test.go
