+++ clients/go/pkg/commands/resolveIncident_command.go
