+++ clients/go/pkg/commands/resolveIncident_command_test.go
