+++ clients/go/pkg/commands/setVariables_command.go
