+++ clients/go/pkg/commands/setVariables_command_test.go
