+++ clients/go/pkg/commands/topology_command.go
