+++ clients/go/pkg/commands/topology_command_test.go
