+++ clients/go/worker/jobWorker.go
@@ -16,7 +16,7 @@
 package worker
 
 import (
-	"github.com/zeebe-io/zeebe/clients/go/commands"
+	"github.com/zeebe-io/zeebe/clients/go/pkg/commands"
 	"github.com/zeebe-io/zeebe/clients/go/pkg/entities"
 	"sync"
 )
