+++ clients/go/worker/jobWorker_builder.go
@@ -16,8 +16,8 @@
 package worker
 
 import (
-    "github.com/zeebe-io/zeebe/clients/go/commands"
     "github.com/zeebe-io/zeebe/clients/go/internal/pb"
+    "github.com/zeebe-io/zeebe/clients/go/pkg/commands"
     "github.com/zeebe-io/zeebe/clients/go/pkg/entities"
     "log"
     "math"
