+++ clients/go/worker/jobWorker_test.go
@@ -19,9 +19,9 @@ import (
     "fmt"
     "github.com/golang/mock/gomock"
     "github.com/golang/protobuf/proto"
-    "github.com/zeebe-io/zeebe/clients/go/commands"
     "github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
     "github.com/zeebe-io/zeebe/clients/go/internal/pb"
+    "github.com/zeebe-io/zeebe/clients/go/pkg/commands"
     "github.com/zeebe-io/zeebe/clients/go/pkg/entities"
     "github.com/zeebe-io/zeebe/clients/go/utils"
     "testing"
