+++ clients/go/zbc/api.go
@@ -18,7 +18,7 @@ package zbc
 import (
 	"time"
 
-	"github.com/zeebe-io/zeebe/clients/go/commands"
+	"github.com/zeebe-io/zeebe/clients/go/pkg/commands"
 	"github.com/zeebe-io/zeebe/clients/go/worker"
 )
